Improved exit function
Improving the exit function by adding the prompt.stop variable and setting the loop exit condition prompt.stop != FALSE.
Each time a valid Exit command is encountered, the prompt.stop variable becomes TRUE and exits the loop.
Also each time an exit command is encountered, the variable s_cmd.exec_stat becomes NON_EXECUTABLE, to warn the executor not to execute this command.
Added s_prompt.token_status so that in case of a single | , the loop is interrupted and the message is printed: Syntax error near unexpected token.
Substitution of the function is_builtin by is_builtin_is_exit.
Improvement in the case that the argument is only CD, it is added as second argument /Users/epresa-c so that the executor takes us to HOME.