add fn_delete_quotes && print_error
The fn_delete_quotes function works in general terms but there are still a few bugs to fix.
The print_error function prints the error if prompt->error_msg != NO_ERROR . For example, if there is a Malloc error, prompt->error_msg = MALLOC_ERROR. This function needs to be added throughout the code to check all MALLOCs.